examples/telegram-bot/requirements.txt
# Telegram Bot Dependencies
aiogram==3.13.1
aiohttp==3.10.5
python-dotenv==1.0.0
websockets==13.1
base58==2.1.1
